+++ libs/devices/api-device/models/src/lib/menu-config/menu-config.ts
